requirements.txt
numpy
pandas
torchmetrics
tqdm
scipy
wandb
gdown
matplotlib
transformers
jupyter
jupyterthemes
biopython
